+++ regex-engine/src/nfagenerator.h
@@ -7,10 +7,10 @@
 
 namespace Regex
 {
-	class NfaParser
+	class NfaGenerator
 	{
 	public:
-		NfaParser(const std::string& infixExpression);
+		NfaGenerator(const std::string& infixExpression);
 
 		struct State
 		{
@@ -28,8 +28,6 @@ namespace Regex
 		NFA getNfa();
 
 	private:
-		std::string addConcatenations(const std::string& rawInfixExpression);
-		std::string infixToPostfix(const std::string& infixExpression);
 		std::shared_ptr<State> createState(const bool isEnd);
 		std::unique_ptr<NFA> generateNfa(const std::string& postfixExpression);
 		void addEpsilonTransition(const std::shared_ptr<State> from,const std::shared_ptr<State> to);
@@ -39,9 +37,6 @@ namespace Regex
 		std::unique_ptr<NFA> concatStates(const std::unique_ptr<NFA> first, const std::unique_ptr<NFA> second);
 		std::unique_ptr<NFA> unionStates(const std::unique_ptr<NFA> first, const std::unique_ptr<NFA> second);
 		std::unique_ptr<NFA> kleeneStarState(const std::unique_ptr<NFA> state);
-		//bool recursiveBacktrackingSearch(const std::string& input, const State& currentState, size_t position, std::vector<std::shared_ptr<State>> visited);
-		bool search(const std::unique_ptr<NFA> nfa, const std::string& word);
-		void addNextState(const std::shared_ptr<State> currentState, std::vector<std::shared_ptr<State>>& nextStates, std::vector<std::shared_ptr<State>>& visitedStates);
 		NFA _nfa;
 	};
 }
