+++ regex-engine/src/parser.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <string>
+
+namespace Regex::Parser 
+{
+	std::string addConcatenations(const std::string& rawInfixExpression);
+	std::string infixToPostfix(const std::string& rawInfixExpression);
+}
