--- regex-engine/src/regexengine.cpp
@@ -1,7 +1,0 @@
-#include <iostream>
-
-int main(int argc, char* argv[])
-{
-	std::cout << "Regex Engine is running..." << std::endl;
-	return 0;
-}
