--- regex-engine/src/regexengine.h
@@ -1,5 +1,0 @@
-//#pragma once
-//
-//int main(int argc, char* argv[])
-//{
-//}
